@@ -25,13 +25,13 @@ module MetaDataHelper
         s = Array(meta_datum.deserialized_value).map do |p|
           next unless p
           #temp# link_to p, p
-          link_to p, search_path(:query => p.fullname)
+          link_to p, media_entries_path(:query => p.fullname)
         end
         s.join('<br />').html_safe
       when "Keyword"
         s = Array(meta_datum.deserialized_value).map do |v|
           next unless v
-          link_to v, search_path(:query => v.to_s)
+          link_to v, media_entries_path(:query => v.to_s)
         end
         s.join(', ').html_safe
       when "Meta::Date"
@@ -300,7 +300,7 @@ module MetaDataHelper
           unless @js_1
             @js_1 = true
             locale = "de-CH"
-            h += javascript_include_tag "jquery/i18n/jquery.ui.datepicker-#{locale}"
+            h += javascript_include_tag "plugins/i18n/jquery.ui.datepicker-#{locale}"
             h += javascript_tag do
               begin
               <<-HERECODE
@@ -316,7 +316,7 @@ module MetaDataHelper
                   $(".datepicker").datepicker(
                     $.extend({
                       showOn: "button",
-                      buttonImage: "/images/icons/calendar.png",
+                      buttonImage: "/assets/icons/calendar.png",
                       buttonImageOnly: true,
                       changeMonth: true,
                       changeYear: true,
@@ -352,8 +352,8 @@ module MetaDataHelper
           h += widget_meta_terms_multiselect(meta_datum, meta_key)
 
         when "Copyright"
-          h += meta_datum.hidden_field :value, :class => "copyright_value"
-###          h += hidden_field_tag field_id, meta_datum.object.value.first, :class => "copyright_value"
+          #old# h += meta_datum.hidden_field :value, :class => "copyright_value"
+          h += hidden_field_tag "#{meta_datum.object_name}[value]", meta_datum.object.value.try(:first), :class => "copyright_value"
 
           @copyright_all ||= Copyright.all # OPTIMIZE
           @copyright_roots ||= Copyright.roots
@@ -383,8 +383,8 @@ module MetaDataHelper
                 var copyrights = {};
                 var custom_copyright_id;
                 $.each(#{@copyright_all.to_json}, function(i,item){
-                  copyrights[item.copyright.id] = item.copyright;
-                  if(item.copyright.is_custom) custom_copyright_id = item.copyright.id; 
+                  copyrights[item.id] = item;
+                  if(item.is_custom) custom_copyright_id = item.id; 
                 });
                 
                 $("select.nested_options, select.options_root").change(function(event){
