@@ -1,29 +1,32 @@
+# -*- encoding : utf-8 -*-
 module Logic
   extend self
 
   def data_for_page(media_entry_ids, current_user)
     media_entries = MediaEntry.find(media_entry_ids)
 
-    editable_ids = current_user.accessible_resource_ids(:edit)
-    managable_ids = current_user.accessible_resource_ids(:manage)
-    favorite_ids = current_user.favorite_ids
-    
-    editable_in_context = editable_ids & media_entry_ids
-    managable_in_context = managable_ids & media_entry_ids
-
     { :pagination => { :current_page => media_entry_ids.current_page,
                        :per_page => media_entry_ids.per_page,
                        :total_entries => media_entry_ids.total_entries,
                        :total_pages => media_entry_ids.total_pages },
-      :entries => media_entries.map do |me|
-                    flags = { :is_private => me.acl?(:view, :only, current_user),
-                              :is_public => me.acl?(:view, :all),
-                              :is_editable => editable_in_context.include?(me.id),
-                              :is_manageable => managable_in_context.include?(me.id),
-                              :is_set => false, # OPTIMIZE
-                              :is_favorite => favorite_ids.include?(me.id) }
-                    me.attributes.merge(me.get_basic_info(current_user)).merge(flags)
-                  end } 
+      :entries => data_for(media_entries, current_user) } 
+  end
+
+  def data_for(media_entries, current_user)
+    editable_ids = current_user.accessible_resource_ids(:edit)
+    managable_ids = current_user.accessible_resource_ids(:manage)
+    favorite_ids = current_user.favorite_ids
+
+    media_entries.map do |me|
+      flags = { :is_private => me.acl?(:view, :only, current_user),
+                :is_public => me.acl?(:view, :all),
+                :is_editable => editable_ids.include?(me.id),
+                :is_manageable => managable_ids.include?(me.id),
+                :can_maybe_browse => !me.meta_data.for_meta_terms.blank?,
+                :is_set => false, # OPTIMIZE
+                :is_favorite => favorite_ids.include?(me.id) }
+      me.attributes.merge(me.get_basic_info(current_user)).merge(flags)
+    end    
   end
 
 ########################################################
@@ -44,7 +47,8 @@ module Logic
       flags = { :is_private => res.acl?(:view, :only, current_user),
                 :is_public => res.acl?(:view, :all),
                 :is_editable => editable_in_context.include?(res.id),
-                :is_manageable => managable_in_context.include?(res.id) }
+                :is_manageable => managable_in_context.include?(res.id),
+                :can_maybe_browse => !res.meta_data.for_meta_terms.blank? }
       all_attributes = res.attributes.merge(res.get_basic_info(current_user)).merge(flags)
       all_attributes.merge!(:url_stub => (resource_type == "MediaEntry") ? "media_entries" : "media_sets")
       all_attributes.merge!(:is_set => resource_type != "MediaEntry")
