@@ -166,7 +166,7 @@ task :clear_cache do
   run "cd #{release_path} && RAILS_ENV=production bundle exec rails runner 'Rails.cache.clear'"
 end
 
-
+before "deploy:symlink", :make_tmp
 after "deploy:symlink", :link_config
 before "configure_sphinx", :link_sphinx
 after "deploy:symlink", :configure_sphinx
@@ -177,7 +177,6 @@ after "deploy:symlink", :record_deploy_info
 #before "migrate_database", :install_gems
 #after "migrate_database", :load_seed_data
 after "migrate_database", :clear_cache
-before "deploy:restart", :make_tmp
 after "install_gems", :stop_sphinx
 after "deploy", :start_sphinx
 after "deploy", "deploy:cleanup"
