@@ -1,3 +1,5 @@
+# coding: UTF-8
+
 Given /^I have set up the world$/ do
   # Set this to a non-JS driver because:
   # 1. The Selenium driver times out during this step
@@ -34,6 +36,7 @@ Given /^I have set up the world$/ do
   # the tests. Therefore we recreate our world in this step.
   Copyright.init
   Permission.init
+
   Meta::Department.fetch_from_ldap
   Meta::Date.parse_all
 
@@ -193,8 +196,8 @@ When /^I create a set titled "([^"]*)"/ do |title|
   create_set(title)
 end
 
-When /^I add the picture "([^"]*)" to the set "([^"]*)"/ do |picture_title, set_title|
-    add_to_set(set_title, picture_title)
+When /^I add the picture "([^"]*)" to the set "([^"]*)" owned by "([^"]*)"/ do |picture_title, set_title, owner|
+    add_to_set(set_title, picture_title, owner)
 end
 
 When /^I toggle the favorite star on the media entry titled "([^"]*)"$/ do |title|
