@@ -1,4 +1,4 @@
-
+# coding: UTF-8
 
 Before do
   # This is run before EVERY scenario. It's necessary to prevent
@@ -109,7 +109,8 @@ def check_media_entry_titled(title)
   # which Capybara can't do)
   make_hidden_items_visible
   entry = find_media_entry_titled(title)
-  cb_icon = entry.find(:css, ".check_box").find("img")
+  #cb_icon = entry.find(:css, ".check_box").find("img")
+  cb_icon = entry.find(:css, "div.check_box")
   #debugger; puts "lala"
   cb_icon.click if (cb_icon[:src] =~ /_on.png$/).nil? # Only click if it's not yet checked
 end
@@ -324,7 +325,7 @@ def upload_some_picture(title = "Untitled")
                                         "Copyright" => 'some dude' })
 
     click_button("Metadaten speichern und weiter…")
-    click_link_or_button("Weiter ohne Hinzufügen zu einem Set")
+    click_link_or_button("Weiter ohne Hinzufügen zu einem Set/Projekt…")
 
     sphinx_reindex
     visit "/"
@@ -343,12 +344,12 @@ end
 # Adds a media entry to a set. Only works if the media entry
 # has a title, so that it shows up under /media_entries. The set
 # also needs a title.
-def add_to_set(set_title = "Untitled Set", picture_title = "Untitled")
+def add_to_set(set_title = "Untitled Set", picture_title = "Untitled", owner = "No one")
   visit "/media_entries"
   click_media_entry_titled(picture_title)
   click_link_or_button("Zu Set/Projekt hinzufügen")
-  select(set_title, :from => "media_set_ids[]")
-  click_link_or_button("Zu ausgewähltem Set hinzufügen")
+  select("#{set_title} (#{owner})", :from => "media_set_ids[]")
+  click_link_or_button("Zu ausgewähltem Set/Projekt hinzufügen…")
   # The set title is displayed on the right-hand side of this page, so we should be able to
   # see it here.
   page.should have_content(set_title)
