@@ -28,17 +28,11 @@ class RedClothMadek
   end
 
   def media_tag(text)
-    # unfortunately having multiple matches in gsub doesn't seem to work, therefore
-    # we fall back to $1 $2
-    text.gsub(/\[\s*media\s*=\s*(\d+)\s*\|\s*([^\]]+)\s*\]/) { 
-      "<a href='/media_entries/#{$1}'>#{h($2.strip)}</a>"
-    }
+    text.gsub(/\[\s*media\s*=\s*(\d+)\s*\|\s*([^\]]+)\s*\]/, "<a href='/media_entries/\\1'>\\2</a>")
   end
 
   def thumbnail_tag(text)
-    text.gsub(/\[\s*thumbnail\s*=\s*(\d+)\s*\|\s*([^\]]+)\s*\]/) { 
-      "<img src='/media_entries/#{$1}/image' title='#{h($2.strip)}'/>"
-    }
+    text.gsub(/\[\s*thumbnail\s*=\s*(\d+)\s*\|\s*([^\]]+)\s*\]/, "<img src='/media_entries/\\1/image' title='\\2'/>")
   end
 
   def include_tag(text)
@@ -54,12 +48,13 @@ class RedClothMadek
         return "Page '#{$1}' does not exist."
       end
     }
+
   end
 
   def video_tag(text)
     text.gsub(/\[\s*video\s*=\s*(\d+)\s*\|\s*([^\]]+)\s*\]/) { 
-      "<video src='/media_entries/#{$1}/image' title='#{h($2.strip)}'>" +
-        "<a href='/media_entries/#{$1}'>(see '#{h($2.strip)}' Video)</a>" +
+      "<video src='/media_entries/\\1/image' title='\\2'>" +
+        "<a href='/media_entries/\\1'>(see '\\2' Video)</a>" +
       "</video>"
     }
   end
